+++ docs/components/button/demos/DemoBasic.vue
@@ -3,7 +3,7 @@
     <gi-button type="add" @click="handleClick"></gi-button>
     <gi-button type="edit"></gi-button>
     <gi-button type="delete"></gi-button>
-    <gi-button type="search"></gi-button>
+    <gi-button type="search" round></gi-button>
     <gi-button type="search" loading></gi-button>
     <gi-button type="reset"></gi-button>
     <gi-button type="upload"></gi-button>