+++ packages/components/button/src/type.ts
@@ -1,15 +1,15 @@
 import type { ButtonProps as ElButtonProps } from 'element-plus'
 
+export type CustomButtonType =
+  | 'add'
+  | 'edit'
+  | 'delete'
+  | 'search'
+  | 'reset'
+  | 'upload'
+  | 'download'
+  | 'print'
+
 export interface ButtonProps extends Partial<Omit<ElButtonProps, 'type'>> {
-  type?:
-    | 'add'
-    | 'edit'
-    | 'delete'
-    | 'search'
-    | 'reset'
-    | 'upload'
-    | 'download'
-    | 'print'
-    | ''
-    | ElButtonProps['type']
+  type?: CustomButtonType | '' | ElButtonProps['type']
 }